Tangible assets rework
