+++ src/main/java/uk/gov/companieshouse/api/accounts/repository/smallfull/TangibleAssetsRepository.java
@@ -1,0 +1,15 @@
+package uk.gov.companieshouse.api.accounts.repository.smallfull;
+
+import org.springframework.stereotype.Repository;
+import uk.gov.companieshouse.api.accounts.enumeration.AccountingNoteType;
+import uk.gov.companieshouse.api.accounts.model.entity.smallfull.notes.tangibleassets.TangibleAssetsEntity;
+import uk.gov.companieshouse.api.accounts.repository.AccountsNoteRepository;
+
+@Repository
+public interface TangibleAssetsRepository extends AccountsNoteRepository<TangibleAssetsEntity> {
+
+    @Override
+    default AccountingNoteType getAccountsNote() {
+        return AccountingNoteType.SMALL_FULL_TANGIBLE_ASSETS;
+    }
+}
