+++ src/main/resources/notes.yml
@@ -9,3 +9,4 @@ notes:
     - intangible-assets
     - creditors-within-one-year
     - creditors-after-more-than-one-year
+    - tangible-assets
